backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy>=1.26.0
pandas>=2.0.3
python-dotenv
requests>=2.31.0
google-generativeai